requirements.txt
asgiref==3.7.2
Django==4.2.3
django-flexible-subscriptions==0.15.1
gunicorn==20.1.0
Pillow==10.0.0
psycopg2-binary==2.9.6
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.7.1
asgiref==3.7.2
boto3==1.27.0
botocore==1.30.0
Collectfast==2.2.0
cssselect2==0.7.0
Django==4.2.3
django-filer==2.2.6
django-flexible-subscriptions==0.15.1
django-js-asset==2.1.0
django-mptt==0.14.0
django-polymorphic==3.1.0
django-s3-folder-storage==0.5
django-storages==1.13.2
easy-thumbnails==2.8.5
gunicorn==20.1.0
jmespath==1.0.1
lxml==4.9.2
Pillow==10.0.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-decouple==3.8
reportlab==4.0.4
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.7.1
Unidecode==1.1.2
urllib3==1.26.16
webencodings==0.5.1
asgiref==3.7.2
boto3==1.27.0
botocore==1.30.0
Collectfast==2.2.0
cssselect2==0.7.0
Django==4.2.3
django-flexible-subscriptions==0.15.1
django-js-asset==2.1.0
django-mptt==0.14.0
django-polymorphic==3.1.0
django-s3-folder-storage==0.5
django-storages==1.13.2
easy-thumbnails==2.8.5
gunicorn==20.1.0
jmespath==1.0.1
lxml==4.9.2
Pillow==10.0.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-decouple==3.8
reportlab==4.0.4
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.7.1
Unidecode==1.1.2
urllib3==1.26.16
webencodings==0.5.1
asgiref==3.7.2
boto3==1.27.0
botocore==1.30.0
Collectfast==2.2.0
crispy-bootstrap5==0.7
cssselect2==0.7.0
Django==4.2.3
django-crispy-forms==2.0
django-flexible-subscriptions==0.15.1
django-js-asset==2.1.0
django-mptt==0.14.0
django-polymorphic==3.1.0
django-s3-folder-storage==0.5
django-storages==1.13.2
easy-thumbnails==2.8.5
gunicorn==20.1.0
jmespath==1.0.1
lxml==4.9.2
Pillow==10.0.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-decouple==3.8
reportlab==4.0.4
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.7.1
Unidecode==1.1.2
urllib3==1.26.16
webencodings==0.5.1
